+++ src/wrath_compat.lua
@@ -1,0 +1,13 @@
+local wrathRequirements = {
+    ["Wistiti-WrathOfOlympus" .. "-" .. "AresWrathBoon"] = {
+        OneFromEachSet = {
+            [2] = { "AresManaBoon", "BloodDropRevengeBoon", "RendBloodDropBoon", gods.GetInternalBoonName("AresBlinkTrailBoon")},
+        }
+    }
+}
+
+if rom.mods["Wistiti-WrathOfOlympus"] and rom.mods["Wistiti-WrathOfOlympus"].config and rom.mods["Wistiti-WrathOfOlympus"].config.enabled then
+    if game.TraitRequirements["Wistiti-WrathOfOlympus" .. "-" .. "AresWrathBoon"] then
+        game.TraitRequirements = DeepMergeUptoDepth(game.TraitRequirements, wrathRequirements, 2)
+    end
+end